fix: address pr review findings (round 1)
