+++ turbo/apps/platform/src/signals/zero-page/__tests__/zero-chat.test.ts
@@ -85,6 +85,7 @@ describe("zero-chat signals", () => {
                 id: "t1",
                 title: null,
                 preview: "Hello",
+                agentComposeId: "mock-compose-id",
                 createdAt: "2026-03-10T00:00:00Z",
                 updatedAt: "2026-03-10T00:00:00Z",
               },
@@ -92,6 +93,7 @@ describe("zero-chat signals", () => {
                 id: "t2",
                 title: null,
                 preview: "World",
+                agentComposeId: "mock-compose-id",
                 createdAt: "2026-03-10T01:00:00Z",
                 updatedAt: "2026-03-10T01:00:00Z",
               },
